change int consts to INTVAL
git-svn-id: https://svn.parrot.org/parrot/trunk@6001 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ imcc/debug.c
@@ -193,7 +193,7 @@ dump_symreg(IMC_Unit * unit)
 		    r->first_ins->index, r->last_ins->index,
 		    r->first_ins->bbindex, r->last_ins->bbindex,
 		    r->set,
-                r->color, r->want_regno, r->score,
+                (int)r->color, r->want_regno, r->score,
                 r->use_count, r->lhs_use_count,
                 r->reg ? r->reg->name : "",
                 r->usage
